+++ src/diagnostics/index.ts
@@ -11,6 +11,7 @@ import { DirectiveParser } from '../directive-parser'
 import { detectMarpDocument } from '../utils'
 import * as deprecatedDollarPrefix from './deprecated-dollar-prefix'
 import * as overloadingGlobalDirective from './overloading-global-directive'
+import * as unknownTheme from './unknown-theme'
 
 export const collection = languages.createDiagnosticCollection('marp-vscode')
 
@@ -20,6 +21,7 @@ const setDiagnostics = lodashDebounce((doc: TextDocument) => {
 
   deprecatedDollarPrefix.register(doc, directiveParser, diagnostics)
   overloadingGlobalDirective.register(doc, directiveParser, diagnostics)
+  unknownTheme.register(doc, directiveParser, diagnostics)
 
   directiveParser.parse(doc)